requirements.txt
dnspython==2.1.0
psutil==5.8.0
pymongo==3.11.3
PyNaCl==1.4.0
python-dotenv==0.18.0
python-dotenv-run==0.1.4
python.py==0.0.3
